backend/requirements.txt
bcrypt==4.1.1
bidict==0.22.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-SocketIO==5.3.6
greenlet==3.0.1
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
python-decouple==3.8
python-engineio==4.8.0
python-socketio==5.10.0
simple-websocket==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
wsproto==1.2.0
